pyproject.toml
[project]
name = "dht22-webui"
version = "1.0.2"
description = "DHT22 Sensor monitoring system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adafruit-circuitpython-dht>=4.0.6",
    "flask>=3.1.0",
    "flask-socketio>=5.4.1",
    "python-engineio>=4.10.1",
    "python-socketio>=5.11.4",
    "RPi.GPIO>=0.7.0",
    "gevent>=24.2.1",
    "gevent-websocket>=0.10.1",
    "pymongo>=4.6.1",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.3.0"]

[tool.ruff]
line-length = 150
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "RUF", # ruff-specific rules
]

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
skip-magic-trailing-comma = false
line-ending = "lf"
